New Upgrade Menu Added!!
now we are able to slowly add upgrades to towers.
the buttons already work on their own but they aren't connected to the tower yet.
the new menu also shows some stats of the tower like the range, the name and rate of fire.
also I've been trying a lot of different art styles for the tiles but haven't had too much luck with it yet
who would had said that drawing space rocks was so hard?